+++ gui/src/renderer/components/Map.tsx
@@ -3,6 +3,7 @@ import { Component, Types, View } from 'reactxp';
 
 import SvgMap from './SvgMap';
 
+// Higher zoom level is more zoomed in
 export enum ZoomLevel {
   high,
   medium,
@@ -84,14 +85,15 @@ export default class Map extends Component<IProps, IState> {
   };
 
   // TODO: Remove zoom level in favor of center + coordinate span
+  // TODO: Zoomlevels below 2.22 makes australia invisible
   private zoomLevel(variant: ZoomLevel) {
     switch (variant) {
-      case ZoomLevel.high:
+      case ZoomLevel.low:
         return 1;
       case ZoomLevel.medium:
-        return 20;
-      case ZoomLevel.low:
-        return 40;
+        return 2.22;
+      case ZoomLevel.high:
+        return 5;
     }
   }