+++ CHANGELOG.md
@@ -35,6 +35,10 @@ Line wrap the file at 100 chars.                                              Th
 - Change project copyright and company name from Amagicom AB to Mullvad VPN AB
 - Only reconnect when settings change if a relevant tunnel protocol is used.
 - Adjust padding of tray icon on Windows and Linux to better match other icons.
+- Change the zoomlevel of the map in the desktop app to make it less zoomed in.
+
+### Removed
+- Remove city/country labels on map in the desktop app.
 
 ### Fixed
 #### Android